Relicense from AGPLv3 to MIT
This is so because **SPYLESS is now divided into backend, and frontend repositories.** The reason for the original license on SPYLESS to be AGPLv3 is because of the frontend. The backend is meant to be very small and have the ability to integrate in all kinds of environments, from a VPN to earn on, to a VPN used in companies, while the original SPYLESS frontend is meant to be, very small too, but also more privacy focused: No hidden tracking, no ADs on the main page. If you'd like to do that, your choice will be only to fully re-do the frontend from scratch.